fixed dyunamic build for xmrigcc and disabled remote pausing

+++ buildscripts/xmrigCC-dynamic-build.sh
@@ -1,20 +1,20 @@
 #!/bin/sh
 
-if [ -z "$(ldd $(which gcc) | grep -i musl)" ]; then
-    export CC="/usr/bin/musl-gcc"
-fi
-export CFLAGS=" -Ofast -Flto "
-export CPPFLAGS="$CFLAGS" CXXFLAGS="$CFLAGS" LDFLAGS="$CFLAGS"
-
-LINKER="ld-musl-x86_64.so.1"
 MYLIBPATH="pl/lib"
 MYLDPATH="$MYLIBPATH/ld.so"
+prevpath=$PWD
 
 ## voidlinux/voidlinux-musl
 cd /
-export MAKEFLAGS=" -j $(nproc) "
-xbps-install -Syu
-xbps-install cmake libuv-devel gcc libmicrohttpd-devel git make
+
+if type apk; then
+    apk add --update alpine-sdk cmake libuv-dev libmicrohttpd-dev coreutils
+    LINKER="ld-musl-x86_64.so.1"
+else
+    xbps-install -Syu
+    xbps-install cmake libuv-devel gcc libmicrohttpd-devel git make
+    LINKER="ld-linux-x86-64.so.2" ## care about underscores
+fi
 rm -rf /xmrigCC
 git clone --depth=1 https://github.com/Bendr0id/xmrigCC
 mkdir xmrigCC/build && cd xmrigCC/build || exit 1
@@ -28,10 +28,22 @@ sed -r '/add_executable\(xmrigDaemon/atarget_link_libraries\(xmrigDaemon -Wl\,--
 ## don't use xmrigDaemon
 sed 's/m_daemonized(false)/m_daemonized(true)/' -i ../src/Options.cpp
 sed -r 's/(kDonateLevel = )([0-9]+)/\10/' -i ../src/donate.h
+patch $PWD/../src/cc/CCClient.cpp ${prevpath}/skipCommand.patch
+
+if [ -z "$(ldd $(which gcc) | grep -i musl)" ]; then
+    export CC=${CC:-"$(uname -m)-linux-musl-gcc"}
+fi
+export MAKEFLAGS=" -j $(nproc) "
+export CFLAGS=" -Ofast -Flto"
+export CPPFLAGS="$CFLAGS" CXXFLAGS="$CFLAGS" LDFLAGS="$CFLAGS"
 cmake ..
 LD_RUN_PATH='$ORIGIN/'"$MYLIBPATH" RPATH='$ORIGIN/'"$MYLIBPATH" make
 mkdir -p "$MYLIBPATH"
 ldd xmrig* | grep "=> /" | awk '{print $3}' | xargs -I '{}' cp -v '{}' "$MYLIBPATH"
-cp "/lib/$LINKER" "$MYLDPATH"
-ln -sr "MYLDPATH" "$MYLIBPATH/$LINKER"
+if [ -f "/lib/$LINKER" ]; then
+        cp "/lib/$LINKER" "$MYLDPATH"
+else
+        cp "/usr/lib/$LINKER" "$MYLDPATH"
+fi
+ln -sr "$MYLDPATH" "$MYLIBPATH/$LINKER"
 tar czf xmrig.tar.gz xmrig* "$MYLIBPATH"
